種々の変更を追加

+++ src/main/java/NagasawaKenji/IsctClassReview/repository/UserRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Short> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     public Optional<User> findByUserName(String userName);
 }
